MLK-20307 gpu: imx: dpu: extdst: Fix get_xval()
commit  92899d70e5c4a0a6a2c2fd609bf52593a6c0147c from
https://source.codeaurora.org/external/imx/linux-imx.git

The function get_xval() returns the C_XVAL and L_XVAL fields of
the CURPIXELCNT and LASTPIXELCNT registers.  They are 16bit and
sit in the low 16bit of the registers.  This patch fixes the way
we mask the register and read the fields out.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>