MLK-20357-1: ASoC: imx-rpmsg: add audio routing for wm8960
commit  5e76ecfef115e2d33cc0d592069e725c0d1ca91e from
https://source.codeaurora.org/external/imx/linux-imx.git

The rpmsg wm8960 codec driver is completed to support
full function, not only the volume control. which cause
an issue that there is no sound when recording, the reason
is that the MIC Bias not enabled, and it should be enabled
through audio routing.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>