drm: xlnx: crtc: Fix max width and height overflow
commit  eeab73d1207d6fc2082776c954eb19fd7290bfbe from
https://github.com/Xilinx/linux-xlnx.git

The maximum resolution supported is shown as -1x-1 in Xorg.log when
trying to run Xorg. This is occurring due to width and height variables
overflowing as they are assigned UINT_MAX instead of INT_MAX.
This patch fixes this by correcting the data types of width and height
variables and correctly assigning INT_MAX as maximum width and height.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Signed-off-by: Saurabh Sengar <saurabh.singh@xilinx.com>
Reviewed-by: Sandip Kothari <sandipk@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>