enetc: Call register_netdev as late as possible, refactoring
This removes the risk of a race at probe, once the netdevice gets
registered. This is a requirement that the upstream maintainers
always verify during review, and point it out if not implemented.
Note that request_irq must be moved to open(), to be able to call
it after register_netdev().

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>
(cherry picked from commit 6bc8a48e5daefa950de182f5b35c106dc554b97f)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>