net: mvpp2: fix tx stop-wake queue for combined cpu-txq map
The tx-stop/wake has constrain - working properly only if
netdev-queue-id (TXQ) selected by core-xmit is absolutely conforms
to the cpu/TXQ mapping.

This is not always correct!
For example, for UDP-Application (APP) the core-xmit always keeps the
same TXQ chosen on APP's start according to a socket-hash.
Even the APP migrates to another CPU the TXQ stays the same.

If TXQ doesn't match to CPU the stop/wake would work incorrectly -
a lot of drops or even could cause stop-without-wakeup.

This patch fixes the constrain
 and handles all CPU<->TXQ combinations.
It saves the TXQ in new per-cpu field "stopped_on_txq_id" and
uses it to wake the correct queue.

NOTE:
The stop/wake/stopped field sizes and if conditions in the patch
Are strictly optimized to keep the same TX performance as before.

Change-Id: I33be0f4c045ff18bdbbdfc0df769efedfb158dd2
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/60974
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>