dma: axi-dmac: terminate early DMA transfers after a partial one
When we get a partial transfer, we should not submit any more segments to
the hardware, as the data it will get will be unused and will just do
unnecessary processing.

During the start of transfer, we check it there was a partial transfer in a
previous completed segment and add the TLAST flag to the next segment that
is about to be submitted.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 70b6b10a3739e93ba3ae987459b972be7384cac2)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>