arm64: dts: renesas: r8a77990: sort subnodes of the soc node
commit a342b3bb8c309826182c28e445b61c349b7e8367 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

This patch sorts the subnodes of the soc node to improve maintainability.
The sort key is the address on the bus with instances of the same
IP block grouped together and sorted alphabetically.

This patch does not include functional changes.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>