can: flexcan: Add warning if prescaler is not uniform
As per can for Auto guidelines the prescaler for data phase and control phase
is desired to be same.
Since this value is calculated by can driver (same for all can controllers),
it is beyond the control of flexcan.
Therefore, just put a warning if the mismatch occurs between two prescaler
values, so that user is able to adjust the bitrate to make the prescaler
uniform.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>
(cherry picked from commit 8d0ee098fd8c79e2df5ec5ba140582712e083189)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>