dma: axi-dmac: Infer synthesis configuration parameters hardware
This patch integrates all the changes from
  git clone https://github.com/analogdevicesinc/linux.git

Some synthesis time configuration parameters of the DMA controller can be
inferred from the hardware itself.

Use this information as it is more reliably than the information specified
in the devicetree which might be outdated if the HDL project got changed.

Deprecate the devicetree properties that can be inferred from the hardware
itself.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>