MLK-20470 drm/imx: dpu: plane: Do full modeset for tile to linear fb switch
commit  2e4a928dd9e868073b2ca962eca816204c549a8b from
https://source.codeaurora.org/external/imx/linux-imx.git

When pixel combiner is used, it turns out that on-the-fly switch from
tile framebuffer to linear framebuffer would cause hardware malfunction
- right half display would be missing and master&slave content shadow
load done event won't come.  Thus, go for a full modeset as a workaround.
Note that we check if the original framebuffer is tile or not for both
primary and overlay planes.  So, this could be over-kill.  The issue was
found when we use 32bit GPU super tile as the original framebuffer on the
primary plane(restart Weston frequently).  However, since we usually
don't do this kind of switch in real graphics, it should be fine.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
(cherry picked from commit 9245bbf650fdeb79ebf869d48b6bd1b43c2bcf3b)
Signed-off-by: Meng Li <Meng.Li@windriver.com>