soc: octeontx2: update SSO HWGRP teardown sequence
Teardown HWGRP LF resources used by the consumer in a sane sequence.
To access the LF registers in PF/VF bar2 from AF space we use the AF
ALIAS space based on the slot id of the PF/VF to which the LF is
attached.
Let SSO teardown continue even when NPA_FETCH is not cleared.

Change-Id: If71bc34dfec539d930c05d330ea8c2eacd041843
Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1768
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>