x86/mm/pat: disable preemption __split_large_page() after spin_lock()
commit 45c6ff4811878e5c1c2ae31303cd95cdc6ae2ab4 from
https://git.kernel.org/pub/scm/linux/kernel/git/rt/linux-rt-devel.git/log/?h=linux-4.19.y-rt

Commit "x86/mm/pat: Disable preemption around __flush_tlb_all()" added a
warning if __flush_tlb_all() is invoked in preemptible context. On !RT
the warning does not trigger because a spin lock is acquired which
disables preemption. On RT the spin lock does not disable preemption and
so the warning is seen.

Disable preemption to avoid the warning __flush_tlb_all().

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[Issue: LIN1018-3362]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>