net: phy: 2.5G mode for fixed PHYs
The CPU port on a DSA switch needs to be connected in fixed-link mode.
This mode uses an emumated PHY that exposes standard registers. However,
multiGigabit PHYs tend tu use standardised C45 registers to exposes
these modes, which aren't supported in swphy.

We could add these modes, and that will probably be done in mainline,
but in a way that is very difficult to backport. It implies reworking
the way we represent link-modes, which impacts all drivers in
net/ethernet and net/phy.

For now, add a way to simulate this feature by hardcoding the
phy_interface_modes that would need such PHYs to be emulated.

Change-Id: I31d82b60bb2fc3bb049b67f85a542d1fae1f1c00
Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>
Reviewed-on: https://sj1git1.cavium.com/6488
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
Tested-by: sa_ip-sw-jenkins
[RH: Original patch taken from marvell 88F3720 board support SDK 10.0-PR2003]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>