net: mvpp2: add PP2.2 GMAC TX FIFO configuration
TX FIFO can be additionally configured in MAC registers,
which is valid when the interface speed is max. 2500Mbps.
Until now, however this setting was done only for PP2.1,
leaving the defaults for the latest NIC revision.

This patch change the TX FIFO minimal threshold as specified
in erratum RM-9725147 (50 for 1G and 136 for 2.5G). Because
the link speed can be changed dynamically move the TX FIFO
to a separate subroutine, which is executed during init and
also in a phylink callback.

Moreover, fix the MVPP2_GMAC_TX_FIFO_MIN_TH_ALL_MASK value and
add updating assertion/deassertion levels of the p2mem_ready_signal,
which indicates readiness for fetching the data from DRAM to the
TX FIFO. Those levels must be set according to the conditions:
TxFifoHiWM > TxFifoLowWM > TxFifoMinTh.

Change-Id: I74be927c7fe66c801b316e6d94b27a2146424a66
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/62044
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1703
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>