arm64: mvebu: add support for Marvell CN 9130 development board
This patch adds basic support for the Marvell CN-9130 Modular
development board. It is based on cn9130 SoC built from one AP-807 and
one CP-115, which is extended via 2xMoCi interface about additional
2xCP-115 (one located on main board and other on board extension).
This setups describes only interfaces related to CP0.

Some of the interfaces found on the board related to CP0:
- UART
- eMMC
- I2C
- SDHCI
- NAND
- SPI NOR
- Ethernet (RGMII, SFI)
- PCIex4
- SATA

Change-Id: I0340e784006847c347b221552efc1cb881274a4e
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/62328
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1709
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06.
Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>