tee: optee: Modify duration of spinlock for list
commit 940b2f6016df46b976c05b4fb6365d93d81a1952 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

R-Car OP-TEE driver acquires spinlock only for the duration of list_del()
when deleting entries from the debug log list.
This may cause that a debug log is left in list until a next RPC is
received.
And list_for_each_entry_safe() is not safe for deleting entries from the
list if the spin lock is released and reacquired during the list iteration.

Therefore, this patch replaces list_for_each_entry_safe() with checking
if list is empty and removing the first entry in each iteration, and
lock for the duration of the sequence.

Signed-off-by: Hiroki Negishi <hiroki.negishi.bx@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>