arm64: mvebu: add support for cn9131 development board
This patch extends support for the Marvell CN-9130 about additional CP-115.
The name of this product is cn9131, the last number indicates how
many external CP-115 are used. There will be one more dts description for
cn9132, which extends the cn9131 about additional CP-115.

Some of the interfaces found on the board related to CP1:
- SPI
- I2C
- SATA
- PCIex2
- USB3.0
- 10G ETH

Change-Id: I460e594ec336f4d0e24650be574f69b1e97feed7
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/2967
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Kostya Porotchkin <Kostya.Porotchkin@cavium.com>
Reviewed-by: Nadav Haklai <Nadav.Haklai@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06.
Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>