arm64: mvebu: add support for cn9132 development board
This patch extends support for the Marvell CN-9131 about additional CP-115.
The name of this product is cn9132, the last number indicates how many
external CP-115 are used.

Some of the interfaces found on the board related to CP2:
- I2C
- SDHCI
- 10G ETH
- PCIex4
- SATA
- USB

Change-Id: Idb2d08ac67f60bafe31f2dc210628c84f485b6ee
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/3170
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Kostya Porotchkin <Kostya.Porotchkin@cavium.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06.
Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>