net: mvpp2: add 10G and 5G support for port1
The CP115 is capable to work with 10G/5G speed on port1. Rework the mvpp2
driver in a way which will allow to configure port1 of CP115 to those
modes.

Introduce a new flag, indicating, whether port can use the XLG MAC.
It can be set only once during initialization.

Change-Id: I9331cbd92af4aadad159de31581328e4d6bdbec1
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/4698
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>