octeontx2-pf: Fix memory leak due to error packets
When HW adds a CQE_RX for a received error packet it also
consumes a buffer, ignoring that will lead to memory leak
and also stalls packet reception as number of buffers in
Aura/Pool will come down with each error pkt.

This patch fixes this by freeing the IOVA back to Aura.

Change-Id: Ib94b70a7ba3c9008a2133c6c809b1ddb0abd4c3e
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/5255
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RH: Original patch taken from marvell 88F3720 board support SDK 10.0-PR2003]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>