octeontx2-pf: do not handle VF link event twice
otx2_open() is shared across both PFs and VFs and internally it may
update link state. Which is fine for PFs which may have already received
link event update. For example during probe.

Things get tricker in case of VFs which are informed about link changes
via different mechanism happening in parallel to otx2_open(). And this
is not necessary a good thing because leads to updating link status
twice.

So to make things behave as expected link handling during otx2_open()
should be limited to PFs only.

Change-Id: I80858cff519cde3702ce4b3a4c18f5cc71f4b4b1
Signed-off-by: Tomasz Duszynski <tduszynski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/8801
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RH: Original patch taken from marvell 88F3720 board support SDK 10.0-PR2003]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>