Revert "drm/gem: drm_gem_dumb_map_offset(): reject dma-buf"
This reverts commit 90378e589192 ("drm/gem: drm_gem_dumb_map_offset(): reject dma-buf").

There are 2 drm devices in i.mx8mq, one is DCSS and the other is
GPU (GC7000 Lite). When starting weston service in userspace, it will
call libgbm to let GC7000 to allocate dma_buf as its buffer, and then
GC7000 will pass a fd which associated to this dma_buf to DCSS.

Now DCSS use the fd passed by GC7000 to find the dma_buf and import it
into its own context. This is implementation of the buffer sharing
between graphic cards.

In upstream commit 90378e589192 ("drm/gem: drm_gem_dumb_map_offset(): reject dma-buf"),
says that it is not allowed that DCSS use the imported dma_buf as a dumb map
buffer.

But the libgbm which provided by NXP uses this unsuggested method and because
it's binary format, we can do nothing but only change kernel source as workaround.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>