ocelot: Filter out ocelot SoC specific PCS config from common path
commit 28757c1efcb8e522753cf5e37f6cfc84d57a70d3 from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

The adjust_link routine should be generic enough to be (re)used by
any SoC that integrates a switch core compatible with the Ocelot
core switch driver.  Currently all configurations are generic except
for the PCS settings that are SoC specific.  Move these out to the
Ocelot SoC/board instance.

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>