kexec: dcss-pll: enable pll output and clear PD down bit to avoid kernel hang
There is a dtg module in DCSS subsystem. Kexec kernel always hang when
accessing dtg registers. This is because that the PLL and powerdomain
for DCSS is disabled by dcss_runtime_suspend in the first kernel. So we
enable clk output and power up VIDEO_PLL2 by calling dcss_pll_enable for
the kexec kernel to avoid kernel hang.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>