net: mvpp2: Restore the BM pool id of the rxq overrode by the MUSDK
When a port is hijacked by the MUSDK, the BM pool id of the rxq will
be set to the one created by the MUSDK, so we should restore the BM
pool id when switching back to kernel. Otherwise the kernel would panic
due to the buffs from MUSDK are used.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>