futex: Make the futex_hash_bucket lock raw
commit f646521aadedab78801c9befe193e2e8a0c99298 from linux-rt-devel

Since commit 1a1fb985f2e2b ("futex: Handle early deadlock return
correctly") we can deadlock while we attempt to acquire the HB lock if
we fail to acquire the lock.
The RT waiter (for the futex lock) is still enqueued and acquiring the
HB lock may build up a lock chain which leads to a deadlock if the owner
of the lock futex-lock holds the HB lock.

Make the hash bucket lock raw so it does not participate in the
lockchain.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[OP: backport of f646521aaded from linux-5.0.y-rt branch of linux-rt-devel]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>