felix: Add packet injection extraction mechanism
commit 1103d6f026c5b17866c2834aeea69d537f5046ee from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

This mechanism allows a faster alternative to send/receive frames
than current Ocelot implementation based on register access.
The new mechanism is based on the Felix capability to allow
frames tagged with a special header to bypass the switch frame
analyzer on xmit. For receive, we use another Felix capability
that enables to forward all frames to the CPU port and have
the frames tagged with 32-byte header from which we can extract
the source port.

Signed-off-by: Catalin Horghidan <catalin.horghidan@nxp.com>
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>