driver: ethernet: add shutdown fundtion for ravb driver on renesas rcar3 platform
When run kexec feature, ethernet device is not able to be opened
the second kernel boots up. Because the ethernet device is still in
running status. So, it is need to do correct release operation before
the second kernel boots up.
Therefore, add shutdown fundtion to stop the enternet device and
release corresponding resource.

Signed-off-by: Meng Li <Meng.Li@windriver.com>