arch: arm64: dts: reduce the size of reserved memory
On renesas rcar platform, reserved 3 blocks memory. The total
size is 704MB, they take up most of low address money. This
will cause kdump feature is not able to work. Because there
is no remaining memory for crash kernel.
So far, there is not any feature need to reserve such big memory.
Therefore, reduce the size of the 2 blocks so that there is enough
memory for the crash kernel and make sure the crash kernel
boots up smoothly.

Signed-off-by: Meng Li <Meng.Li@windriver.com>