net: thunderx: Adjust CQE_RX drop levels for better performance
With the current RX RED/DROP levels of 192/184 for CQE_RX, when
packet incoming rate is high LLC is getting polluted resulting
in caches misses and higher latency in packet processing. This
slows down the whole process and performance loss. Hence reduced
reduced the levels to 224/216 (ie for a CQ size of 1024, Rx pkts
will be red dropped or dropped when unused CQE are less than 128
/160 respectively)

Change-Id: I1a19e4309644861e05400d78b625291d1dd53355
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/12340
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RH: Original patch taken from marvell 88F3720 board support SDK 10.0-PR2003]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>