net: octeontx2-af: Reset the rvu af mxi-x table in case of kexec/kdump
The RVU AF MXI-X table region's base is configured by firmware, the HW
interprets it as IOVA, so we will create a IOMMU mapping for this
physical address and write IOVA into the RVU_AF_MSIXTR_BASE register.
But the physical address configured by the firmware may be needed by
the second kernel in case of kexec/kdump, so we add the shutdown hook
and panic hook for this driver to restore the physical address
configured by firmware.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>