dt-bindings: PCI: designware: Remove the num-lanes from Required properties
commit 0fc8620887937c2dea1724ea2e49b411738b90b2 from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

The num-lanes is not a mandatory property, e.g. on FSL
Layerscape SoCs, the PCIe link training is completed
automatically base on the selected SerDes protocol, it
doesn't need the num-lanes to set-up the link width.

It is previously in both Required and Optional properties,
let's remove it from the Required properties.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>