octeontx2-af: Compatible with the old firmware version
In commit 64ca66734d05 ("octeontx2-af: Move to rvu_fwdata version 1."),
we have changed the RVU_FWDATA_VERSION to 1 to match with the latest
firmware version, but this will break on the board which is still using
the old version firmware. So add the support to be compatible with the
old version firmware.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>