enetc: Fix hang with MDIO WA
commit df5ae7bb69d7b4acceedf90c5462c1b7bb4588da from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

The ENETC Tx clean-up flow ends up calling kmem_cache_free_bulk which
cannot be called with IRQs off.  The previous version of the MDIO WA
was keeping IRQs off over this call, this is now fixed.
Use a single lock in the enetc MSI handler and disable interrupts before
disabling preemption in MDIO accessors, following the model in the generic
code.

Fixes: 6907d6d1d1cc ("enetc: WA for MDIO register access issue")

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>