arm64: dts: ls1028a: fix dwc pci over smmu
commit 106285f8e4f65b88e6b7889c377955f1ffd7411a from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

In order for the dwc controller to work with SMMU it needs the
bootloader to fixup it's iommu-map property. In the current
implementation to bootloader will not perform the fixup if the
property is not already in the device tree with dummy fields.
Add it to fix DWC PCI over SMMU.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>