irqchip: imx-irqsteer: add clock operations during probe
Some platforms like i.MX8QXP has clock control for this module
irqsteer, which as an interrupt controller cooperates with GIC,
gathers interrupts from the system and can be used to steer them
to one of the output bits (non A-series core).

Need to add clock operations and enable channel to IRQs during
initialization to make sure the driver is working properly.

This will fix the kdump hung issue when calling machine_kexec_mask_interrupts()
to mask the related interrupt but hung due to the irqsteer is in a
messy state.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>