drivers: net: flexcan: fix using wrong interrupt flag register
This is to fix the issue no response after flexcan interrupts came.

The interrupt flag 1 & 2 defines the flags for the 32 message
buffer interrupts for MB64 to MB0, by default it should be for flag 1
but not flag 2 causing irq handler is not correct. Here fix it,
which was introduced by wrong patch picking and integrating.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>